+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Enable Filetote on "move" command <https://github.com/gtronset/beets-filetote/pull/124>
+
 ## [0.4.4] - 2023-11-23
 
 ### Fixed